+++ src/components/FinalResult/FinalResult.js
@@ -2,12 +2,12 @@ import React from 'react';
 import './FinalResult.css';
 import scoreBackground from '../../assets/images/score-background.svg';
 import { Button } from '../Button/Button';
-import { getFinalScore } from '../getFinalScore';
-import { getAlbumResults } from '../getAlbumResults';
+import { getFinalScore } from '../../utils/getFinalScore';
+import { getAlbumResults } from '../../utils/getAlbumResults';
 import { ShareButton } from '../ShareButton/ShareButton';
 import twitterLogo from '../../assets/images/twitter-logo.svg';
 import facebookLogo from '../../assets/images/facebook-logo.svg';
-import { getResultDescription } from '../getResultDescription';
+import { getResultDescription } from '../../utils/getResultDescription';
 
 const twitterBaseURL = 'https://twitter.com/intent/tweet';
 const facebookBaseURL =
