+++ src/utils/getCurrentRoundResult.js
