+++ src/utils/getFinalScore.js
