+++ src/utils/resultText.js
