+++ src/utils/setNextRoundArtistOptions.js
